Created VirtualInputJoystick: A type of joystick that hooks into the InputManager event system to push positively/negatively when two buttons are pushed.
Moved WillSnap to Virtualinputjoystick: It's a rather more situational option that can be invalid with some potential children of Joystick, so I just moved it to an actual implementation of the joystick.

Made isValuePositive and isValueNegative in Joystick: Might as well have some simple boolean shortcuts for these things. it'll be inline so it'd resolve to the same thing as if it wasn't a function anyway.

Made clampToMinMaxSize static: it doesn't require anything from the individual Joysticks, so why have it local?