Created Framerate.h to store static and semi-static data related to the timing of the program without needing to include the entire GameEngine class.
VirtualInputJoystick::GetPushForce can now accept a boolean value to determine if you want the function to return the push force as a negative or positive integer, with the function returning a positive value by default.

Attempted to ratelimit updates of keys to prevent flickery inputs. We'll see how that goes but I'm pretty sure I need to do something more elaborate to truly fix that.

Player no longer prints the x value used to determine how far to move it each frame, that only existed for debugging purposes anyway.